Changed Setup API
The plugin expects that sonarlint-language-server will be started with
the flag -stdio and that the list of analyzers JARs are provided with
the option -analyzers.

Additionally, the user should provide now a list of filetypes instead of
a map of filetypes and analyzers. IMHO this should make the
configuration more understandable.

Further, the README.md provides some additional version hints.